Reorganizing AP register definitions
